+++ subgrid_radiation/sun_position_array/rays_comp.h
@@ -9,47 +9,47 @@ void sw_dir_cor_comp(
     int dem_dim_0, int dem_dim_1,
     float* vert_grid_in,
     int dem_dim_in_0, int dem_dim_in_1,
-    float* sun_pos,
+    double* sun_pos,
     int dim_sun_0, int dim_sun_1,
     float* sw_dir_cor,
     int pixel_per_gc,
     int offset_gc,
     uint8_t* mask,
-    float dist_search,
+    double dist_search,
     char* geom_type,
-    float sw_dir_cor_max,
-    float ang_max);
+    double sw_dir_cor_max,
+    double ang_max);
 
 void sw_dir_cor_comp_coherent(
     float* vert_grid,
     int dem_dim_0, int dem_dim_1,
     float* vert_grid_in,
     int dem_dim_in_0, int dem_dim_in_1,
-    float* sun_pos,
+    double* sun_pos,
     int dim_sun_0, int dim_sun_1,
     float* sw_dir_cor,
     int pixel_per_gc,
     int offset_gc,
     uint8_t* mask,
-    float dist_search,
+    double dist_search,
     char* geom_type,
-    float sw_dir_cor_max,
-    float ang_max);
+    double sw_dir_cor_max,
+    double ang_max);
 
 void sw_dir_cor_comp_coherent_rp8(
     float* vert_grid,
     int dem_dim_0, int dem_dim_1,
     float* vert_grid_in,
     int dem_dim_in_0, int dem_dim_in_1,
-    float* sun_pos,
+    double* sun_pos,
     int dim_sun_0, int dim_sun_1,
     float* sw_dir_cor,
     int pixel_per_gc,
     int offset_gc,
     uint8_t* mask,
-    float dist_search,
+    double dist_search,
     char* geom_type,
-    float sw_dir_cor_max,
-    float ang_max);
+    double sw_dir_cor_max,
+    double ang_max);
 
 #endif
